✨ feat: add custom file icon layer provider
Implements a new `CustomFileIconLayerProvider` to display a 
dynamic icon layer based on the size of the editor group. 
This enhances the visual representation of editor groups 
by providing a counter on the file icons, improving user 
experience and organization. Removes unused code related 
to the editor tab title provider for clarity.